Remove old implementation of SQL parser
