+++ integrations/airflow/marquez_airflow/extractors/sql/__init__.py
@@ -9,3 +9,5 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from marquez_airflow.extractors.sql.parser import SqlParser, SqlMeta
